requirements.txt
web3==4.2.0
eth-abi==1.0.0
eth-account==0.2.2
eth-bloom==1.0.0
eth-hash==0.1.2
eth-keyfile==0.5.1
eth-keys==0.2.0b3
eth-rlp==0.1.2
eth-tester==0.1.0b21
eth-utils==1.0.3
ethereum==1.6.1
pyyaml==3.12
dpath==1.4.2
py-solc==2.1.0
eth-testrpc==1.3.3
timeout-decorator==0.4.0
six==1.11.0
requests==2.18.4
structlog==18.1.0
python-json-logger==0.1.8
boto3==1.5.31
tendo==0.2.8
apsw==3.9.2.post1
pytest==3.4.2
pytest-cov==2.5.1
rlp==0.6.0
watchtower==0.5.3
psutil==5.4.5
pycodestyle==2.4.0
